fix: first integrated terminal proces being killed
When we open a new terminal, we get only a shellProcessId, not a
processId. We therefore have to assume that the first process we get is
our program, but for instance *nix nvm runs an npm command when a shell
is first opened. This causes us to think the debug session is over
before it's begun, and terminate any further processes.

This gets us _better_ behavior by not assuming the session is over when
the first process exits, the downside being that the session will remain
open until the user closes the shell.

I think we need a fix in vscode to send us the `processId` in the
initial request to fix this correctly.